+++ R/utils.R
@@ -38,37 +38,45 @@ presentation_commit <- function(commit) {
     "\n",
     paste("- Date:", commit$when),
     "\n",
-    paste("- Content:\n"),
+    paste("- Message content:"),
+    "\n",
     "```",
     "\n",
-    commit$message,
+    paste(commit$message),
     "\n",
-    "```"
+    "```",
+    "\n",
+    paste("- Related patterns:\n"),
+    # "```{block, type='commit-content'}",
+    # "\n",
+    paste(commit$message_link)
+    # "\n",
+    # "```"
   )
 }
 
 #' Write presentation for each commit
 #'
 #' @param commits table of commits of one pattern as issued from nest_commits_by_pattern
-#' @param name_section character. Name of the section
-#' @param pattern_name Name of the pattern selected
+#' @param pattern.type character. Name of the section
+#' @param pattern.content Name of the pattern selected
 #' @param link_pattern Name of the pattern selected transformed as slug
 #'
 #' @importFrom purrr map_chr
 #'
 
-each_commit <- function(commits, pattern_name, link_pattern, name_section) {
+each_commit <- function(commits, pattern.content, link_pattern, pattern.type) {
   c(
     # One Pattern Section
-    paste0("## ", to_singular(name_section), ": ",
-           pattern_name, "{#", link_pattern, "}"),
+    paste0("## ", to_singular(pattern.type), ": ",
+           pattern.content, "{#", link_pattern, "}"),
     # Commits
     map_chr(1:nrow(commits), ~presentation_commit(commits[.x, ]))
   )
   # repo <- fake_repo()
   # one_pattern_commits <- nest_commits_by_pattern(repo, pattern = "ticket[[:digit:]]+",
   # ref = "master", silent = TRUE)[1,]
-  # each_commit(one_pattern_commits, name_section = "Ticket")
+  # each_commit(one_pattern_commits, pattern.type = "Ticket")
 }
 
 #' To singular
@@ -79,43 +87,47 @@ to_singular <- function(x) {
 
 #' Presentation for each pattern
 #'
-#' @param pattern vector of regex patterns
-#' @param name_section names for pattern section
-#' @param repo path to the repo
-#' @param ref branch of the repo
+#' @param nest_commits commits as nested with nest_commits_by_pattern
+#' @param pattern.type Character name of the pattern to filter
 #'
-#' @importFrom dplyr tibble mutate bind_rows
-#' @importFrom purrr map_chr
+#' @importFrom dplyr tibble mutate bind_rows filter
+#' @importFrom purrr map_chr pmap
 #' @export
 #' @examples
 #' repo <- fake_repo()
-#' each_pattern("ticket[[:digit:]]+", "Ticket", repo)
+#' res_commits <- nest_commits_by_pattern(
+#'   repo,
+#'   pattern = c("Ticket" = "ticket[[:digit:]]+"),
+#'   ref = "master", silent = TRUE
+#' )
+#' each_pattern(res_commits, "Ticket")
 
-each_pattern <- function(pattern, name_section, repo, ref = "master") {
+each_pattern <- function(nest_commits, pattern.type) {
+
+  filter_commits <- nest_commits %>%
+    filter(pattern.type == !!pattern.type)
 
   # Create text of each commit by pattern
-  res_commits <- nest_commits_by_pattern(
-    repo,
-    pattern = pattern,
-    ref = ref, silent = TRUE
-  ) %>%
+  res_commits <- filter_commits %>%
     mutate(
-      text = map(data, ~each_commit(.x, pattern_name, link_pattern, name_section))
+      text = pmap(list(data, pattern.content, link_pattern, pattern.type),
+                  function(commits, pattern.content, link_pattern, pattern.type)
+                    each_commit(commits, pattern.content, link_pattern, pattern.type))
     )
 
   # Create summary table
   summary_table <- tibble(
-    pattern_name = res_commits$pattern_name,
+    pattern.content = res_commits$pattern.content,
     Links = map_chr(
       res_commits$data,
       ~paste0("[", .x$summary, "](#", .x$link_commit, ")") %>%
         paste(collapse = ", "))
   )
-  names(summary_table)[1] <- name_section
+  names(summary_table)[1] <- pattern.type
 
   # Create text for section
   text_section <- paste0(
-    paste0("# Section: ", name_section, "\n"),
+    paste0("# Section: ", pattern.type, "\n"),
     "## Summary table\n\n",
     paste(knitr::kable(summary_table, format = "markdown"),
           collapse = "\n")
@@ -123,8 +135,8 @@ each_pattern <- function(pattern, name_section, repo, ref = "master") {
 
   res <- bind_rows(
     tibble(
-      pattern_name = name_section,
-      link_pattern = clean_link(name_section),
+      pattern.content = pattern.type,
+      link_pattern = clean_link(pattern.type),
       text = list(text_section)
     ),
     res_commits
@@ -136,35 +148,97 @@ each_pattern <- function(pattern, name_section, repo, ref = "master") {
 #' Nest all commits by each pattern
 #'
 #' @inheritParams get_commits_pattern
-#' @importFrom dplyr filter mutate arrange group_by
-#' @importFrom stringr str_extract_all
+#' @importFrom dplyr filter mutate arrange group_by if_else
+#' @importFrom dplyr ungroup distinct select summarise left_join
+#' @importFrom stringi stri_extract_all
 #' @importFrom tidyr nest
 #'
 #' @export
 #' @examples
 #' repo <- fake_repo()
-#' nest_commits_by_pattern(repo)
-nest_commits_by_pattern <- function(repo, pattern = "#[[:digit:]]+",
+#' nest_commits_by_pattern(repo, pattern = )
+
+nest_commits_by_pattern <- function(repo,
+                                    pattern = c("Issues" = "#[[:digit:]]+"),
                                     ref = "master", silent = TRUE) {
 
-  get_commits_pattern(repo, pattern = pattern,
-                      ref = ref, silent = silent) %>%
+  res <- get_commits_pattern(repo, pattern = pattern,
+                             ref = ref, silent = silent) %>%
     mutate(
       pattern_numeric =
-        str_extract_all(pattern, "([[:digit:]]+)",
-                        simplify = FALSE) %>%
+        stri_extract_all(pattern.content, regex = "([[:digit:]]+)",
+                         simplify = FALSE) %>%
         unlist() %>%
         as.numeric(),
-      pattern_name = pattern,
-      link_pattern = clean_link(pattern_name),
-      link_commit = paste0(link_pattern, "-", sha)) %>%
-    arrange(pattern_numeric, pattern_name, order) %>%
-    # arrange(pattern_name, order) %>%
+      link_pattern = clean_link(paste(pattern.type, pattern.content)),
+      link_commit = paste0(link_pattern, "-", sha))
+
+  correspondance <- res %>%
+    select(sha, pattern.type, pattern.content, link_pattern) %>%
+    distinct() %>%
+    # filter(!is.na(pattern.content)) %>%
+    mutate(pattern.content = if_else(
+      is.na(pattern.content),
+      paste("No related", tolower(pattern.type)), pattern.content)) %>%
+    mutate(
+      # pattern.content.regex = paste0("(?<!\\[)", pattern.content),
+      # text_inter = paste0("-[", pattern.content, "]-"),
+      text_link = paste0("[", pattern.content, "](#", link_pattern, ")")
+    ) %>%
+    # Longest chain to smallest
+    # arrange(desc(nchar(pattern.content))) %>%
+    group_by(sha, pattern.type) %>%
+    summarise(
+      message_link.type = paste(text_link, collapse = ", ")
+    ) %>%
+    group_by(sha) %>%
+    summarise(
+      message_link = paste(paste0("  + ", pattern.type, ": ", message_link.type),
+                        collapse = "  \n")
+    ) %>%
+    ungroup()
+
+# c("# text #1\n issue#1\n#2  #145  #") %>%
+  # c("toto") %>%
+  # stringi::stri_replace_all_regex(
+  #   pattern = correspondance$pattern.content.regex,
+  #   replacement = correspondance$text_inter,
+  #   vectorize_all = FALSE) %>%
+  # stringi::stri_extract_all_fixed(
+  #   pattern = correspondance$text_inter,
+  #   omit_no_match = TRUE
+  # ) %>% unlist() %>% unique() %>%
+  # # Replace with url
+  # stringi::stri_replace_all_fixed(
+  #   pattern = correspondance$text_inter,
+  #   replacement = correspondance$text_link,
+  #   vectorize_all = FALSE)# %>%
+  # stringi::stri_replace_all_regex(replacement = "\\\\#", pattern = "^(#)") %>%
+  # stringi::stri_replace_all_regex(replacement = "\\\n\\\n", pattern = "\\n")
+
+  res %>%
+    # rowwise() %>%
+    left_join(correspondance, by = "sha") %>%
+    # mutate(
+    #   message_link =
+    #     # Replace with url
+    #     stringi::stri_replace_all_regex(
+    #       message,
+    #       pattern = correspondance$pattern.content.regex,
+    #       replacement = correspondance$text_link,
+    #       vectorize_all = FALSE
+    #       ) %>%
+    #     stringi::stri_replace_all_regex(replacement = "\\\\#", pattern = "^(#)") %>%
+    #     # Go to newline
+    #     stringi::stri_replace_all_regex(replacement = "  \\\n", pattern = "\\n")
+    # ) %>%
+    arrange(pattern.type, pattern_numeric, pattern.content, order) %>%
     # If is.na => New part
-    filter(!is.na(pattern_name)) %>%
-    group_by(pattern_numeric, pattern_name, link_pattern) %>%
-    # group_by(pattern_name, link_pattern) %>%
-    nest()
+    # filter(!is.na(pattern.content)) %>%
+    mutate(pattern.content = if_else(is.na(pattern.content), paste("No related", tolower(pattern.type)), pattern.content)) %>%
+    group_by(pattern.type, pattern_numeric, pattern.content, link_pattern) %>%
+    nest() %>%
+    ungroup()
 }
 
 #' Clean link
@@ -172,11 +246,10 @@ nest_commits_by_pattern <- function(repo, pattern = "#[[:digit:]]+",
 
 clean_link <- function(x) {
   paste0("a", x) %>% # First must be a text
-  gsub("^_|^ | $|_$|#|@|`|\\\\|/", "", .) %>%
-  gsub("[[:space:]]+|_+", "-", .) %>%
-  gsub("(-)+", "-", .) %>%
-  gsub("^(-)|(-)$", "", .) %>%
-  tolower(.)
-  # clean_link(" -Text With # \\\\ \ / _characters_ ")
+    gsub("^_|^ | $|_$|#|@|`|\\\\|/|\\[|\\]|\\(|\\)|\\{|\\}|\"|\\+|:", "", .) %>%
+    gsub("[[:space:]]+|_+", "-", .) %>%
+    gsub("(-)+", "-", .) %>%
+    gsub("^(-)|(-)$", "", .) %>%
+    tolower(.)
 }