+++ NEWS.md
@@ -1,3 +1,8 @@
 # gitdown 0.0.1
 
+Breaking changes
+* `names_section` parameter is replaced by a named vector in `pattern`
+
+Other changes
+* Added summary tables for each section
 * Added a `NEWS.md` file to track changes to the package.